+++ internal/federation/discovery_test.go
@@ -15,7 +15,7 @@ import (
 func createTestCAPIClusterWithDetails(name, namespace string, opts ...clusterOption) *unstructured.Unstructured {
 	cluster := &unstructured.Unstructured{
 		Object: map[string]interface{}{
-			"apiVersion": "cluster.x-k8s.io/v1beta1",
+			"apiVersion": "cluster.x-k8s.io/v1beta2",
 			"kind":       "Cluster",
 			"metadata": map[string]interface{}{
 				"name":              name,