+++ internal/federation/types.go
@@ -167,7 +167,7 @@ var (
 	// CAPIClusterGVR is the GroupVersionResource for CAPI Cluster objects.
 	CAPIClusterGVR = schema.GroupVersionResource{
 		Group:    "cluster.x-k8s.io",
-		Version:  "v1beta1",
+		Version:  "v1beta2",
 		Resource: "clusters",
 	}
 )